Fix MOTD 404
Fix the bug. Instead of using `this.$root.papi` which has an API prefix baked into it (and thus attempts to hit a non-existent MOTD endpoint), I reverted back to creating a one-off api with no prefix that succeeds.

Updated the tests to adapt to the corrected behavior.

go mod tidy.