Fix unexpected error popup
GET /api/case/tasks previously was returning 404 without a body. During the rewrite, I started returning the error. This was causing a error popup in the browser. Not returning the error results in the request returning 200 with a blank response as the UI expects.

TODO: more diverse error codes. Right now when the store returns an error we assume it can only mean one thing: 404.

BaseHandler is completely gone. The last remnants in it's files were moved to host.go and host_test.go.