Quicklinks added to Configuration
Mostly static HTML additions to include some quicklinks on the config page before a setting is selected. Needed to add some functionality for the page to support updating it's params when routing to the already loaded component.

i18n strings match the order of how they appear in the lists.