Async Import and Reporting Results
After the uploaded file has been copied to disk, respond with Accepted (202) back to the client and do the transfer and importing from a goroutine. The dashboard URL generated after import is now returned back to the user over the websocket. The UI provides the link in an info popup.