Hunt FilterToggles tweaks
When hunting for query, we add a new query string value for controlling optional filter toggles. Specifically we want the new "exclude soc logs" toggle to be disabled.

A new FilterToggle is being added to soc.json (see branch of same name in securityonion repo). It has a new label and some new behavior.