Rewrite to use Case page
Adding to a Case from hunt now involves a dialog that only lists MRU Cases + the option to create a new case.

Specifying an existing case will open a page to that case and leave you on the Add Observable page with the value prefilled.

Specifying "Create New" creates a new case with the default title and description then leaves you on the Add Observable page with the value prefilled.

New or existing, this approach does not save the Observable until you click the 'Add' button on the cases page. This gives the user time to enter metadata.

Removes unnecessary changes from previous approach.