Salt Options
The salt relay has a default timeout of 30s. This is not long enough to process the default file size limit of pcap/evtx imports (25MB). Instead of enlarging the timeout for all salt operations, salt now respects a variable on the passed in context. If that variable holds a longer timeout than the currently configured one, the new timeout is used instead.

Sending files and importing files now request a 2 minute timeout.