Make configurable
The longer timeout is now configurable. If the long timeout is set to be shorter than the normal timeout, the normal timeout is used instead.

Also, the option is applied inside of the `SendFile` and `Import` methods so every sent file and import benefits from the new timeout instead of just the calls in gridmembershandler.go.