WIP: Track Options, Override UpdatedAt
Track is a field for both Suppress and Threshold type overrides but it contains different enum values depending on which of those two types, basically Threshold doesn't have `by_either`.

When saving a detection, update the UpdatedAt field for an override when the old detection doesn't have an override with the exact same values regardless of order.