WIP: Bulk Actions, Removal of Date Range from Detections Search
Detections can now be selected for bulk actions. The Select All checkbox toggles between selecting the entire current page and clearing all selection. If the current page is selected, a prompt to select ALL of the results becomes available. Currently the only actions that can be taken are the enabling or disabling of selected detections.

If picking individual IDs, then those IDs are sent up to the API to be modified. If ALL are selected, then the query is sent and detections matching the query will be modified.

A few component tests were added to hunt.test.js testing the SelectAll logic. Cypress tests will be added later.

UpdateDetectionField has been updated to update a detection field (i.e. without having the entire object) faster. Also should no longer be limited to one field although the functionality is not tested or implemented.

The ElasticDetectionstore now gets an ElasticSearch client passed into it so actions don't have to be run against Eventstore.

An error was fixed in the suricata parser. Better care is taken around modifying settings files that may be empty.

Settings' validator that used to check for Jinja now checks for opening AND closing jinja tags to prevent accidental false positives such as the suricata rule that contained `{%` in a URL.