Allow and Deny Regexes per Detection Engine
(Also includes CSS fixes for Detection DB details being better centered below the detail boxes.)

ElastAlert, Suricata, and Strelka now support `allowRegex` and `denyRegex` config options. These regexes are applied to community rules during the import process. The regex is applied to the entire rule body, one rule at a time. The rules are optional and case sensitive.

ElastAlert and Suricata regexes are applied before parsing the rules. Since Yara files tend to contain multiple rules, the rules must be parsed first and then filtered.

If an allowRegex is specified, the rule must match to be included. If a denyRegex is specified, the rule must NOT match to be included. If both regexes are specified and match, the rule will be denied. If both regexes are specified but do not match, the rule will be denied. Rules filtered out will not be stored in ElasticSearch or SOC.

Yara imports now are only including for the immediate following rule body and not all remaining rules in the file. Yara imports are considered part of the rule body when testing the regexes.

Changing any regex for any engine will not have an impact on already imported rules. However, existing rules may have their updates denied by these regexes. Due to the regexes being defined in soc.json, changing their values requires restarting the SOC service.

Updated tests.