New Create Detection Form
Slimmer create detection form now only asks for language, license, and the rule source. PublicId, Severity, and Title are extracted from the rule source.

Converting detections now works by sending the body of the rule to be converted. This allows us to convert rules that haven't been saved yet.

Creating a detection now checks that the PublicId isn't taken before allowing the create to happen.

Fixed a bug when retrieving a detection that doesn't have a ruleset.

Todo: tests