New Alert Quick Action: Tune Detection
When you click to toggle the QuickAction menu, an API call is sent out to look up the Detection's ID using the PublicID. When the ID comes back, the new Tune Detection option will be visible in the menu. This option opens up that Detection's page to the Tuning tab.