Bug Fixes and Tweaks
Don't show the title on the Detection Create page.

Added a few classes and IDs to help tests.

When extracting references, check URLs with Regex so we don't turn plain text into a hyperlink.

Fixed a bug where extracting logic wouldn't extract anything if the unindent process didn't unindent anything.

When extracting details from sigma rules, extract description too if present.

Better logic around extracting yara logic. The `strings:` section is useful but not required. If not found, fallback to looking for `condition:` which is required.