Moved override update logic from the detectionhandler to the detectionstore. The handler should check for all the ways the request will fail. When everything succeeds, the store should do all the manipulation. This also prevents a redundant call to the DB for the existing detection.
Removed "tautologies" that the linter was pointing out.

Moved another validator from Casestore to Detectionstore for validating a Detection's tags.

Fixed a bug when validating the ruleset of a detection.

When validating a detection, check the language against the engine to be sure it's a valid pairing.

DoesTemplateExist now considers any 2XX successful instead of just 200.

Removed a couple redundant auth checks.

Updated Casestore tests to have unique test names with Detectionstore and to fix tests that fail from an introduced change. ValidateStringArray was assuming the label was "Tags" even though a label was passed in, but the value was "tags".

Detectionstore tests.

TODO: More Detectionstore tests.