Add Version Number to External JS Filenames
Added the version number to jsyaml a few commits ago. But that was incomplete:

1) I changed the filename and added the import in test_common, but I forgot to update the script tag in index.html.
2) I should have made the same change to the external lz string library.

jsyaml now has the version number in the filename everywhere it's referenced.

The LZString library has been given the same treatment. The 1.5.0 was added to the filename, the script tag reference was updated, and the LZString library has been added to global in test_common in line with the others.