Removed PublicID from the Detection Source tab now that we have it in the details on the right.
Cleanup the received detection when loading data.

Include loading shade when deleting a detection. The server has to sync so even if the deletion is quick, the work may not be done quickly.

Renamed functions in detectionhandler.go for consistency. I'm opting to use the verb of what's actually happening rather than the HTTP verb.

Fix an issue when deleting a detection. Forgot to call disableCrossClusterIndexes.