Detection Author and ClickToEdit Class
When creating a custom detection, the author field will be filled in with the current user's First Name and Last Name and falling back to the user's Email. If a detection is a community rule, it'll list the ruleset as the author. If the detection is not a community rule, this author field is displayed.

Override fields now have the clicktoedit class that highlights and changes the cursor on mouseover. The class is removed while the element is being edited.