Better Behavior around Ctrl+C
When modules are shutting down, their sleeps are now interrupted in the detection engines. Also, more checks have been added so the engines can exit during any step of processing.

Also has a small tweak to the clicktoedit styling added so the styles are only active when not editing those controls.

Put the checkboxes back in overrides.