Manually Sync Detections From UI
In Detections, under the Advanced Options, is a new dropdown and a pair of buttons. Selecting a detection engine and clicking one of the buttons will kick off the engine processing community updates. A Differential Update only checks for new updates, a Full Update will reprocess all existing rules.

If a sync for that engine is already ongoing, the user is signed up to receive the status notification just the same as if the user had kicked off that import.

If buttons are clicked multiple times for the same engine, only one sync happens (whichever type was clicked first).

Manual syncs will notify the user in the UI with the results of the sync when they finish. Automatic syncs will not report.

Updating logging around beginning and ending of syncs.

Some stale or commented out code has been removed as well.