Details Tweaks
Extracted Author, Created Time, and Updated Time now all show up unconditionally.

SOC Author: if the rule is a community rule, this will be the community source that the rule came from. If the rule is not a community rule, it'll be the author who saved the custom detection (specifically their SOC user's First Name + Last Name or if no name is present Email)