Internationalized Detection Author
"SOC Import" is now "__SOC_IMPORT__" and is converted back to "SOC Import" using i18n. Updated the UI to localize this value in the detections page unexpanded rows, the detections page expanded rows, the SOC details on the detection detail page, the expanded history rows in the history tab of the detection detail page.

Updated tests to reflect value change.