Timer State
All engines now write out the unix timestamp of the last time a sync was successfully completed. When the service restarts, these state files are used to schedule the next sync. If a state file is missing, the corresponding engine will wait 20 mins and then sync for the first time. If the state file can't be read on startup, the service will attempt to delete it and things will get back on track next time it's written to.