Enable QuickActions in Detections
The main thing preventing the quickAction menu from appearing before was `this.actions` not receiving anything from the parameters. Being undefined, it threw an error when iterating over it and the thrown error prevented the quickAction menu from appearing. By allowing the menu to appear even when there are no actions, the menu can appear and seems to function.