Disable After Failed Sync on Update
When updating a single detection, if the sync fails then the handler will attempt to disable and resync to maintain integrity. A special return code is used that the client watches for and displays a warning when it happens.

Needed to regenerate mocks after giving up on a particular refactor approach.