Cleanup for Cypress Tests
The way the visual glitch was fixed used an async function. Cypress and async functions don't always play well together, occasionally causing the function to be called twice. Handling the async with the Promise approach solves the visual glitch while allowing cypress to behave more predictably.

Some casing changes in the UI tripped up the backend, the backend now normalizes the data it assumed was normalized.

Moved the check for 206 when saving a detection down so we are either showing the warning or the successful popup, not both.