Sigma Overrides
A custom filter is a small bit of YAML. Updated the UI to use textareas and pre tags to support the multiline nature of the value.

Updated how we apply overrides to meet downstream expectations. This change should apply both when the rule is being converted at the "Detection Source" tab as well as when it's being sync'd.

Updated tests.