Duplication Fixes, Improvements
Cleaned up how we process a detection returned from the API so we always do everything we need to do. This fixes a weird issue with duplication where after duplicating you could hit cancel on the Detection Source tab and the page would convert back to the previous detection you had duplicated.

`convertDetection` no longer accepts parameters. It uses `this.detect` which is fine for both new and pre-existing detection conversion.

Fixed a case insensitive comparison that was hiding the Convert button on the New Detection screen.