Update Suricata Engine
I tried to merge how we handle syncing Local vs Community detections but the processes are just different enough that it's easier to work on as 2 different functions.

Cleaned up a lot of the inner work of the mega functions so they're orchestrated better and a little less oversized.

Needs testing before merging.