Refactored UpdateDetection
Some updates weren't happening so I took a closer look at UpdateDetection. We want to be able to update most fields of any detection for one reason or another. Even community rules get updated versions. Some protections were getting in the way so they were moved up to the handler so they only applied to updates a user is making. This change invalidated some existing tests.

Changing from ETOPEN to ETPRO (and similar changes to where community rules are pulled from) will, on next sync, remove rules that are no longer present from the new import places.

The check that asks the question "Has a detection changed enough to warrant updating?" now checks if the ruleset has changed or if there are Overrides.

When updating repos, any folders that existed before updating that weren't listed for an update are removed. It's assumed we always try to update every repo we're getting community rules from.

Organized the util functions, added tests.