Small but Important Bugs
When validating a sigma rule client-side, first check if it has a publicID before checking that the ID matches the one in the rule. If the detection doesn't have a public ID then the rule is probably a duplicate in which the public ID will be extracted on next save to give the user a chance to change the publicID.

Better error messaging around UPDATING a detection to have a publicID that already exists. This behavior used to be unique to CREATING detections.