Smarter YARA Imports
YARA files often contain multiple rules. If any of those rules need to import anything, they're imported once at the top of the file. Now, as SOC parses files it retains those imports and re-adds them to rules as it individualizes them making sure to only add the imports that are actually referenced.