Sanitize Strings
Gridmembershandler.go had a custom implemented sanitize function, replaced it with the sanitize library of choice we use else where in the project.

PublicIds are used in a few places to name files being written to disk so some extra care was taken to be sure we validate them more stringently and sanitize them before use. They are, after all, externally provided content.

More test cases for validateId to ensure file paths or even the characters used in file paths aren't acceptable regardless of slash direction. This is used upstream of where we use PublicIds for file names.