Respect User Input
When I added the new logic for a suricata detection to only update the enabled/disabled pillars if the sid is present in those pillars I forgot to add a way to breakthrough when it's present. Now we mark on the context if the change originated from a user action (incoming web requests are user actions, syncing community rules is not) and the suricata engine respects that flag.

Also found a TODO in bulk update about permissions. Added the permission check.

Updated tests.