Error Tracking and Retry Timing
When updating/creating detections in ElasticSearch during a sync, if a configured number of consecutive errors takes place then the sync should be considered a failure. This assumes that ElasticSearch is the cause and needs more time and less pressure in order to attempt to fix itself.

When a sync fails, we don't wait to wait another sync cycle before trying again in case it was an intermittent failure. Now, when a sync doesn't complete successfully, the engines will wait `communityRulesImportErrorSeconds` (ideally a shorter duration than `communityRulesImportFrequencySeconds`) before trying again. This retry will be marked as a forced sync in hopes of powering through fixing things.