Populate User Details
Added a set of calls to populateUserDetails so detection history users are filled in correctly.

Because this is one of the first instances of SOC creating a model, I modified populateUserDetails to watch for the new system user id of 00000000-0000-0000-0000-000000000000 and report is as "System" when seen.

Note that the system user id was also changed in the backend this commit so any existing detections will continue to report N/A as they have a userId of "agent"