Smarter Strelka Sync
When strelka is deleting all the old enabled rules before writing the new enabled rules, it now deletes any files ending in the `.compiled` extension.

When it comes to the compilation step, only run the python script if there are any enabled detections. Otherwise log that the step is skipped.

Both of these better handle the edge case of having 0 enabled detections.