Strelka
Sometimes after writing to ElasticSearch, we attempt to read back the document but get back an "Object not found" error. Something is causing the ES refresh to take a long time and we should stop syncing detections until ES fixes itself.

When this happens, we now fail the currently executing sync and save the publicId. On next sync, we check if we can find the detection by the publicId. If we can't, we fail the sync again and try again on next sync. If we succeed in retrieving the detection, we clear the flag and continue syncing like normal.