Detections UI Flag
Give every engine an EngineState and for now put the status of the IntegrityCheck on it. The integrity check will report no problem until a problem is found. The UI will watch these flags and show an exclaimation point next to Detections in the navbar if any of these values is false.