Integrity Check Plumbing
We should be able to start a full sync without notifying the result to the user. Refactored InterruptSleep into InterruptSync. Updated the detectionhandler and all 3 engines accordingly.

The IntegrityChecker function is engine agnostic and controls the timing of when crucial elements happen. It can be controlled through it's parameters. `thread`, `interrupt`, and `isModuleRunning` are only used to interrupt the sleep before graceful module shutdown. `isCheckerRunning` allows for the checker to be interrupted and disabled such as during a sync and re-enabled after.

The IntegrityChecker has been added to ElastAlert who's IntegrityCheck function will be implemented first.