Strelka Compilation Report Integrity Check
Strelka's compile_yara.py script now spits out a report that we can use to nail down exactly what rules are deployed vs attempted to deploy (compilation errors).

The integrity check reads this script, verifies the hash in the report matches the hash of the rules.compiled file, then checks that every enabled Detection's publicId is listed in the report.

Also, I noticed a bug while strelka was syncing where if an update to ES failed, we accidentally overwrote the detection object and continued to reference it resulting in a nil exception. Ignoring the returned value from the UpdateDetection call fixes the issue.