Bug Fixes
Integrity Check was coming back bad because we took the address of a stack variable rather than move it to the heap. Fixed in all engines.

Suricata was blindly overwriting the enabled state of imported community rules based on the previous community rule state and not respecting the "if the user hasn't explicitly modified the detection then roll with the enabled or disabled state as defined by the community" requirement which was causing the integrity check to fail as these rules would be enabled but not deployed.

Fixed a crash when converting a strelka rule without an author to a detection.

Fixed the logging of lastSyncStatus. Being a pointer, every time but the first time it would print an address (the first time it's always nil) instead of a value. Now we account for the tri-state and log the proper state in all engines.