YARA Parsing Fix
Expanded YARA parser to treat regex as a string.

Removed YaraRule.String() as it was not recreating rules byte for byte. Removed test that accompanied it. Updated DuplicateDetection to stop using it. YARA rules now retain their exact source during parsing resulting in better integrity of the source that's deployed.

A regex is now used to update the identifier in the source.

Cleaned up code.