Strelka PublicId Change
Strelka used to generate a public ID based on a hash of the rule identifier. To make this easier elsewhere in the system, the new publicId will be the identifier.

We used to sanitize the UUID publicIds before saving the rules to disk but sanitization resulted in duplicate names. Instead, the strelka parser has been updated to be stricter about unacceptable names.

ValidateDetection is updated to call the new ValidatePublicID which contains a regex that should pass any publicID for any engine, including strelka's new potentially 128 character long publicIds.

Tune Detections in Alerts now expects a column called `rule.uuid` in the groupby clauses. This ID is used to look up the SOC ID the same way for all detections regardless of engine.

Updated tests.