Button visibility and unsaved changes to source
The Update/Cancel buttons under the Detection Source now only appear when the source is modified. Sigma's Convert button still remains visible when appropriate.

Changing the source, not saving it using the provided Update button, but causing the detection to save some other way now results in an Unsaved Changes dialog appearing asking you if you want to save the source before continuing. Yes saves, No reverts the source.

Also fixed a bug where canceling your source changes linked the current model to the saved `origDetect` in a way that shouldn't have happened.

If saving a detection fails, the status indicator is reverted back to the original value to give the user a better experience.