dateAwareSort for v-data-tables
Some cypress tests fail because the newly added item wasn't added to the top or bottom of a table as expected. This is because those tables aren't sorting their dates as dates, but as strings. This new date aware sort fixes that. It's not actually aware that the data contains a date, it flags off the field being sorted to determine if it should be converted to date before sorting. Currently sorting by "createTime" and "updateTime" will treat their values as dates, everything else is converted to strings.

The new sort has been applied to Case Observables, Case History, and Detection History.

Hunt seems to use a different format for it's timestamps such that when sorted as a string it will sort it chronologically. Leaving that table untouched for now.

Added test. Cypress tests already assume sorts work so they don't need updating.