Check for Community Rules
Server side now checks if any included detection is a community detection before hand off to the async process. If 1 detection is a community rule, the request is thrown back with a 400 and an error the UI can show without deleting a single rule.

Also, show proper count when every detection is selected and localize the number (i.e. include thousands separator commas).