Only disallow bulk action w/community rules on Delete
When adding BulkDelete, a check for the involvement of community rules was added since you can't delete community rules. However the check will fail Bulk Enables/Disables too. This is causing Cypress tests to fail.

In the sigma sync process, we had an extra bit of logic to copy over all the detections of existing community rules. This was removed as line 1044 accomplishes the task as well.

Added a coverage folder to the gitignore. Jest coverage reports shouldn't be checked in.