Slimmed down Detections Configs
Detections was only using about half of the HuntingParameters so instead of embedding them, I added the fields we need. Updated the Verify and added a test inspired by how Hunt's parameters are verified and tested.

While looking for errors in a cypress test, I noticed a log statement that was unconditionally an error log when it didn't always log a problem. Fixed it.