Suricata Custom Rulesets
New customRulesets config option allowing the specification of download targets or local files for import.

During a sync, the custom rulesets rules are parsed and added deterministically before deduplication. If a ruleset file or target-file can't be read, the sync is aborted otherwise the `delete unreferenced community rules` part of the sync would remove all of them.

Tests.