When duplicating a suricata rule, ensure the msg option's value is quoted
The msg option must be quoted for suricata to accept the rule. Now we use strconv.Quote to put the quotes we removed back.

However strconv.Quote plays with escape sequences because it tries to convert Target Strings to the source code string that when printed would print the Target String. Because we're still dealing with source (suricata source code) we must unescape everything in the string to get the desired result.