New Query Param for Advanced Config
Specifying `a=1` on the query string will now cause the config to load with Advanced already toggled to true.

The nextTick is necessary because on mount, the browser will call processRouterParameters and then loadData. The component is also watching `advanced` and will call refreshTree if it changes.

What happens is that onMounted will call processRouterParameters and set advanced which queues up a refreshTree in the event stack. Then execution exits processRouterParameters and enters loadData. loadData is async so the first half of the function up to the request executes. Then the queued up refreshTree executes. refreshTree checks autoExpand, expands, and then unsets autoExpand. When refreshTree finishes, control is eventually returned to the async part of loadData because the request came back. The second half of loadData calls refreshTree, but this time autoExpand is false so the tree isn't expanded.

With the $nextTick, I queue up the call to refreshTree same as before when I modify advanced in processRouterParameters, then nextTick queues up a function to set autoExpand = true again but after that refreshTree event in the queue. Now when loadData calls refreshTree, autoExpand is set, and the tree expands as expected. We only need to expand on nextTick if we load the page as advanced.