SyncScheduler - ElastAlert
Rough pass on the SyncSchedular, it handles all the timing of when a sync happens. Only ElastAlert is using it at this time.

Updated existing tests with consistent variable naming, added forgotten calls to ctrl.Finish, and to reflect changes to ElastAlertEngine's struct.

Added missing copyright comments to a few files.