Add optional logger to IntegrityCheck
IntegrityCheck now supports adaptive logging. If a logger is passed in the logger gets a new field (intCheckId) and is used. If a nil logger is passed in a new one is built with default fields. This allows IntegrityChecks after a Sync to have the syncId AND an intCheckId.

The IntegrityCheck report log has been condensed. The two lists of IDs enabledButNotDeployed and deployedButNotEnabled are now reported on the same log line as the pass/fail text. Failures are logged as warnings now.