Suricata + Scrolling
The bulk indexer wasn't playing well with our approach to pagination. I believe the concurrency of the import resulted in multiple documents with the same @timestamp value and we were losing some of them at the seams of pages. When expecting 50425 documents back, the paging approach was returning 50415 or 50418. Somewhere between 2 and 20 documents weren't coming back but WERE in ES. This scrolling approach is more thorough.

Suricata is using the new Bulk Indexer

All tests are passing.