Fix Updates
Refactored updates to match the expected syntax. In a create you send up the raw doc, in an update you send it up wrapped in `{"doc":<doc>}`.

Turns out we were missing some errors. OnFailure seems to only provide an error object if there was an error making the request. The response object needs to be checked when there's no err. Also if the bulk indexer runs into a request level error, none of the work items get their OnFailure called, instead the BulkIndexer's OnError gets called.

Creating a bulk indexer now takes a logger which is used to print any errors that happen with the bulk indexer.

Now logging bulk inserter stats as Debug. Updated tests to account for this.

Fixed a potential race condition if multiple bulk written items fail. Prevented a panic that happens when there are multiple writes to the errMap.

Fixed a few places where PrepareForSave was clobbering the docID from previous detections.

It turns out that the backend was not preventing you from deleting a community rule. The frontend would disable the Delete button, but the backend would do it if the request was made. Useful for testing, but needed fixing. Updated with custom error message.