Refactor BulkUpdate to use BulkIndexer
Although the bulk update is incredibly fast, we still need to sync 1 to 3 engines after it's done so the handler is still split into 2 parts: a sync part that verifies you're not trying to delete a community rule, and an async part that makes the changes, does the necessary syncs, and broadcasts the results.

The UI now displays a loading circle during the call to start a bulk action.

Fixed a bug in elastalert sync that was preventing audit entries from getting into ES.

Removed the UpdateDetectionField function from the DetectionStore interface. It was only used in Bulk Updates and has been completely replaced.

TruncateMap was made generic.