Cleanup
Refactored EventScrollCriteria to only the necessary parts. The criteria are used to make a large request with no metrics, no date ranges, no sorting, no search after.

The converter cleans up to just the query and the size which is hardcoded at the ES search limit. A few tests have been added.