Fixed Scrolling
The error was happening because in environments with a large number of shards the scrollID that's generated is larger than 4096. When put in the URL in subsequent scroll requests, ES would return a URL too long error. By putting the scrollID in the body, we don't have to worry about this.

Also fixed an issue where ClearScroll was wiping out any scrolling errors. Now we just log as a warning because dangling scrolls are a self-solving problem. A failure to close the scroll should not slow down whatever work is being done but the problem should be logged in case it's a symptom of a bigger problem.