Fix for Strelka Rule Duplication
The titleUpdater regex didn't account for comments between the identifier and the opening curly bracket of the rule. Updated the regex so that's no longer a problem.

When building a new title for the rule, check against ES to see if the new publicId is taken, if so increment a tailing number up to 10 times before giving up.

Updated the UI to report any errors during the duplication request.