No More Overshadowing Error
When making the initial search, if an error was returned we put it in a variable with a limited, shadowing scope. By the time the function has returned, that error has gone out of scope and is not reported. This caused an empty result set to appear to not be an error.