Respond to IndexNotFound Exception
On first run, the so-detection index won't be created until the first detection is inserted. This means all calls to GetAllDetections were erroring out. A recent refactor ensured any errors were definitely being captured and as a result this exception was bubbling up and causing all syncs to fail.

Now when we run into this case, we log that we noticed the error but we return an empty result set without any error. This will allow syncs to continue and insert their first batch, afterwards everything returns to normal.

Added a test for this use case.