Overrides Added to Detection History
Because we're putting a data table in a data table, we're disabling all the hover effects for a cleaner UI.

Updated dateAwareSort to be aware of new date fields and to properly sort non-string fields (such as isEnabled).

In the tuning table, I pass the `type` fields through i18n so they're the proper case.

Duplicated the logic in findHistoryChange into findOverrideHistoryChange but refactored for the simpler case. We don't need to highlight changes while ignoring source code changes. Added tests.