Majority of Work
Each engine now keeps track of the AI Summaries generated for all the rules of that engine. When the module starts and during Syncs, an engine will update the AI repo and reload the YAML file.

The UI will show the AI Summary if it is present and marked as reviewed. Otherwise, the UI falls back to the extracted description.

On the detection getter, call the new MergeAuxilleryData function on the proper engine so that this AI info is present on the detection when requested. This is the only endpoint that returns AI data, searches will not contain these fields.

readAiSummary is mocked to make progress. It will be implemented soon.

All engines should be configured to put the AiRepo in the same location but this location should NOT be the same as Sigma's rule repos. This would present a problem if somebody didn't use SO's Sigma rules.