Added Rule Validator Regex
Because new detections let you specify the language and not the engine, some things needed tweaking to apply to both new and existing detections.

New regexes were added to validate rules as well as some custom logic in elastalert and suricata to hunt for things tricker than regexes can find.

ElastAlert requires an id field be present, checks that the publicId from the DB matches the publicId in the rule, and lastly checks to make sure that there's some logic present by looking for the detection field and ensuring it's not empty.

Suricata requires there to be no newlines, that the sid is specified, and that the sid matches the DB publicId.

Strelka requires that the identifier is specified, and that the `condition:` header be present. Checking for condition content is a slippery slope because comments might be there complicating things and I don't want to risk a false positive error.