Ai Summary Stale Indicator
Since the Ai Summary doc has the MD5 fingerprints of every rule to track when a rule changes and needs a new summary, I can use that on the backend to know if the summary applies to the latest rule or not. Because the rules don't change drastically, we'll continue to show the summary, but with an additional line that it might be old.

Included a way for users to turn off AI summaries on a per engine basis.