PullRepo Branch Reference, Ctrl+C while Unmarshalling
The branch is also necessary when pulling from a non-standard branch repo. Updated IOManager.PullRepo to accept a branch.

Because of the number of Suricata rules, parsing the AI Summaries yaml file can take 30+ seconds. Other YAML libraries weren't reading any faster. readAiSummary was refactored to watch the value of isRunning and abandon the unmarshalling (safe, no side effects) early if the module terminates before unmarshalling is finished. This allows for a good response time to quit on ctrl+c.

Renamed mio to iom in tests for consistency.